@@ -1,6 +1,8 @@
 /* eslint-disable import/no-unresolved, import/extensions */
-import React, { PropTypes, PureComponent } from 'react';
-import { View, Platform, Animated, Easing, StyleSheet } from 'react-native';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
+import { Platform, Animated, Easing, StyleSheet } from 'react-native';
+import { ViewPropTypes } from '../utils';
 /* eslint-enable import/no-unresolved, import/extensions */
 
 import BottomNavigationAction from './BottomNavigationAction.react';
@@ -22,10 +24,11 @@ const propTypes = {
     * Inline style of bottom navigation
     */
     style: PropTypes.shape({
-        container: View.propTypes.style,
+        container: ViewPropTypes.style,
     }),
 };
 const defaultProps = {
+    active: null,
     hidden: false,
     style: {},
 };
