@@ -1,6 +1,7 @@
 /* eslint-disable import/no-unresolved, import/extensions */
 import { View } from 'react-native';
-import React, { PureComponent, PropTypes } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 /* eslint-enable import/no-unresolved, import/extensions */
 import RippleFeedback from '../RippleFeedback';
 
@@ -10,6 +11,8 @@ const propTypes = {
     style: PropTypes.object,
 };
 const defaultProps = {
+    children: null,
+    onPress: null,
     style: {},
 };
 const contextTypes = {
@@ -50,7 +53,7 @@ class Card extends PureComponent {
 
         if (onPress) {
             return (
-                <RippleFeedback onPress={onPress}>
+                <RippleFeedback onPress={onPress} pointerEvents="box-only">
                     {content}
                 </RippleFeedback>
             );
@@ -58,7 +61,6 @@ class Card extends PureComponent {
 
         return content;
     }
-
 }
 
 Card.propTypes = propTypes;
