@@ -1,10 +1,11 @@
 /* eslint-disable import/no-unresolved, import/extensions */
 import { View } from 'react-native';
-import React, { PureComponent, PropTypes } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 /* eslint-enable import/no-unresolved, import/extensions */
 
 const propTypes = {
-    children: PropTypes.node,
+    children: PropTypes.node.isRequired,
 };
 
 class Container extends PureComponent {
