@@ -1,6 +1,7 @@
 /* eslint-disable import/no-unresolved, import/extensions */
 import { View } from 'react-native';
-import React, { PureComponent, PropTypes } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 /* eslint-enable import/no-unresolved, import/extensions */
 
 const propTypes = {
@@ -36,7 +37,6 @@ class DialogFooter extends PureComponent {
             </View>
         );
     }
-
 }
 
 DialogFooter.propTypes = propTypes;
