@@ -1,5 +1,6 @@
 /* eslint-disable import/no-unresolved, import/extensions */
-import React, { PureComponent, PropTypes } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 import { View } from 'react-native';
 /* eslint-enable import/no-unresolved, import/extensions */
 import Button from '../Button';
@@ -46,7 +47,7 @@ class DialogDefaultActions extends PureComponent {
 
         return (
             <View style={styles.defaultActionsContainer}>
-                {actions.map(action =>
+                {actions.map(action => (
                     <Button
                         key={action}
                         primary
@@ -59,7 +60,7 @@ class DialogDefaultActions extends PureComponent {
                             },
                         }}
                     />
-                )}
+                ))}
             </View>
         );
     }
