@@ -1,5 +1,6 @@
 /* eslint-disable import/no-unresolved, import/extensions */
-import React, { PureComponent, PropTypes } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 import { ScrollView } from 'react-native';
 /* eslint-enable import/no-unresolved, import/extensions */
 import Container from '../Container';
