@@ -1,5 +1,6 @@
 /* eslint-disable import/no-unresolved, import/extensions */
-import React, { PureComponent, PropTypes } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 import { View, TouchableWithoutFeedback } from 'react-native';
 /* eslint-enable import/no-unresolved, import/extensions */
 import ListItem from '../ListItem';
@@ -13,6 +14,9 @@ const propTypes = {
     footer: PropTypes.object,
 };
 const defaultProps = {
+    avatar: null,
+    accounts: null,
+    footer: null,
     style: {},
 };
 const contextTypes = {
@@ -85,12 +89,12 @@ class HeaderAcount extends PureComponent {
                                 // TODO: slice of accounts
                                 // add more soficticated slice when there will be lots of accounts
                                 accounts &&
-                                accounts.slice(0, 3).map((account, index) =>
-                                    <TouchableWithoutFeedback key={index} onPress={account.onPress}>
+                                accounts.slice(0, 3).map(account => (
+                                    <TouchableWithoutFeedback onPress={account.onPress}>
                                         <View style={[styles.inactiveAvatarContainer]}>
                                             {account.avatar}
                                         </View>
-                                    </TouchableWithoutFeedback>,
+                                    </TouchableWithoutFeedback>),
                                 )
                             }
                         </View>
