@@ -1,6 +1,7 @@
 /* eslint-disable import/no-unresolved, import/extensions */
 import VectorIcon from 'react-native-vector-icons/MaterialIcons';
-import React, { PureComponent, PropTypes } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 /* eslint-enable import/no-unresolved, import/extensions */
 
 const propTypes = {
@@ -9,7 +10,11 @@ const propTypes = {
     size: PropTypes.number,
     color: PropTypes.string,
 };
-const defaultProps = { };
+const defaultProps = {
+    size: null,
+    color: null,
+    style: null,
+};
 const contextTypes = {
     uiTheme: PropTypes.object.isRequired,
 };
