@@ -1,5 +1,6 @@
 /* eslint-disable import/no-unresolved, import/extensions */
-import React, { PureComponent, PropTypes } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 import { Animated, TouchableWithoutFeedback, Text, TextInput, Easing, Platform } from 'react-native';
 /* eslint-enable import/no-unresolved, import/extensions */
 
@@ -13,6 +14,9 @@ const propTypes = {
     onSearchTextChange: PropTypes.func.isRequired,
 };
 const defaultProps = {
+    onPress: null,
+    centerElement: null,
+    searchable: null,
     style: {},
 };
 const contextTypes = {
@@ -102,7 +106,7 @@ class CenterElement extends PureComponent {
         //    object again to this instance, search text and isSearchActive will be still set
         let content = null;
 
-        if (isSearchActive) {
+        if (searchable && isSearchActive) {
             content = (
                 <TextInput
                     ref={(ref) => { this.searchFieldRef = ref; }}
