@@ -1,23 +1,21 @@
 /* eslint-disable import/no-unresolved, import/extensions */
-import React, { PureComponent, PropTypes } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 import {
     Animated,
     Easing,
     Platform,
-    BackAndroid as DeprecatedBackAndroid,
-    BackHandler,
     StyleSheet,
     Text,
     View,
 } from 'react-native';
+import { ViewPropTypes, BackAndroid } from '../utils';
 /* eslint-enable import/no-unresolved, import/extensions */
 import LeftElement from './LeftElement.react';
 import CenterElement from './CenterElement.react';
 import RightElement from './RightElement.react';
 import IconToggle from '../IconToggle';
 import isFunction from '../utils/isFunction';
-
-const BackAndroid = BackHandler || DeprecatedBackAndroid;
 
 const propTypes = {
     /**
@@ -65,12 +63,12 @@ const propTypes = {
     * You can overide any style for the component via this prop
     */
     style: PropTypes.shape({
-        container: View.propTypes.style,
-        leftElementContainer: View.propTypes.style,
+        container: ViewPropTypes.style,
+        leftElementContainer: ViewPropTypes.style,
         leftElement: IconToggle.propTypes.style,
-        centerElementContainer: View.propTypes.style,
+        centerElementContainer: ViewPropTypes.style,
         titleText: Text.propTypes.style,
-        rightElementContainer: View.propTypes.style,
+        rightElementContainer: ViewPropTypes.style,
         rightElement: IconToggle.propTypes.style,
     }),
     /**
@@ -145,20 +143,23 @@ const defaultProps = {
     elevation: 4, // TODO: probably useless, elevation is defined in getTheme function
     style: {},
     hidden: false,
+    isSearchActive: false,
+    onRightElementPress: null,
+    rightElement: null,
+    searchable: null,
+    onPress: null,
+    centerElement: null,
+    leftElement: null,
+    onLeftElementPress: null,
+    size: 24,
 };
 const contextTypes = {
     uiTheme: PropTypes.object.isRequired,
 };
 
-const EMPTY_BACK_BUTTON_LISTENER = { remove: () => {} };
-const getBackButtonListener = (callback, isSearchActive) => {
-    // if search is active by default we need to listen back button
-    if (isSearchActive) {
-        return BackAndroid.addEventListener('closeRequested', callback);
-    }
+const getBackButtonListener = callback =>
+    BackAndroid.addEventListener('hardwareBackPress', callback);
 
-    return EMPTY_BACK_BUTTON_LISTENER;
-};
 // const isSearchable = props => (props.searchable && props.isSearchActive) || false;
 // const getIsSearchActive = (props, state) => (props.searchable && state.isSearchActive) || false;
 
@@ -178,10 +179,8 @@ class Toolbar extends PureComponent {
         super(props);
 
         const isSearchActive = props.isSearchActive || false;
-        this.backButtonListener = getBackButtonListener(
-                                        this.onSearchCloseRequested,
-                                        isSearchActive,
-                                    );
+        this.backButtonListener = isSearchActive ?
+            getBackButtonListener(this.onSearchCloseRequested) : null;
 
         this.state = {
             // indicates if searc is activated
@@ -200,6 +199,12 @@ class Toolbar extends PureComponent {
         };
     }
     componentWillReceiveProps(nextProps) {
+        // if search is active and we clicked on the results which does not allow search
+        // then close the previous search.
+        if (this.state.isSearchActive && !nextProps.searchable) {
+            this.onSearchCloseRequested();
+        }
+
         // there should be also posibility to change search through props, so we need to check
         // props first and then we should check state if we need to change search state
         if (this.props.isSearchActive !== nextProps.isSearchActive) {
@@ -236,7 +241,7 @@ class Toolbar extends PureComponent {
             this.state.defaultScaleValue.setValue(0.01);
             this.setState({ order: 'searchFirst' });
             // on android it's typical that back button closes search input on toolbar
-            this.backButtonListener = getBackButtonListener(this.onSearchCloseRequested, true);
+            this.backButtonListener = getBackButtonListener(this.onSearchCloseRequested);
         });
     }
     onSearchPressed = () => {
@@ -273,8 +278,6 @@ class Toolbar extends PureComponent {
             // default scale set up back to "hidden" value
             this.state.searchScaleValue.setValue(0.01);
             this.setState({ order: 'defaultFirst' });
-            // on android it's typical that back button closes search input on toolbar
-            this.backButtonListener = getBackButtonListener(this.onSearchCloseRequested, false);
 
             this.onSearchClosed();
         });
@@ -284,7 +287,9 @@ class Toolbar extends PureComponent {
     onSearchClosed = () => {
         const { searchable } = this.props;
 
-        this.backButtonListener.remove();
+        if (this.backButtonListener) {
+            this.backButtonListener.remove();
+        }
 
         if (searchable && isFunction(searchable.onSearchClosed)) {
             searchable.onSearchClosed();
@@ -451,7 +456,6 @@ class Toolbar extends PureComponent {
             </Animated.View>
         );
     }
-
 }
 
 Toolbar.propTypes = propTypes;
